chore: #227 폴더 변경
